Adjust to changed directory structure
