+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V1/SafeLineHandle.cs
@@ -7,31 +7,53 @@ using LibgpiodV1 = Interop.LibgpiodV1;
 namespace System.Device.Gpio.Libgpiod.V1;
 
 /// <summary>
-/// Pointer to a pin.
+/// Pointer to a pin (Not a real SafeLineHandle, because we need to align its finalization with the owning object)
 /// </summary>
-internal class SafeLineHandle : SafeHandle
+internal sealed class SafeLineHandle : IDisposable
 {
-    public PinMode PinMode { get; set; }
-
+    private IntPtr _handle;
     public SafeLineHandle()
-        : base(IntPtr.Zero, true)
     {
+        _handle = IntPtr.Zero;
     }
 
-    protected override bool ReleaseHandle()
+    public SafeLineHandle(IntPtr handle)
     {
-        // Contrary to intuition, this does not invalidate the handle (see comment on declaration)
-        LibgpiodV1.gpiod_line_release(handle);
-        return true;
+        _handle = handle;
+        PinMode = PinMode.Input;
+    }
+
+    public PinMode PinMode { get; set; }
+
+    public IntPtr Handle
+    {
+        get
+        {
+            return _handle;
+        }
+        set
+        {
+            _handle = value;
+        }
     }
 
     /// <summary>
-    /// Release the lock on the line handle. <see cref="LibgpiodV1.gpiod_line_release"/>
+    /// Release the lock on the line handle. <see cref="Interop.libgpiod.gpiod_line_release"/>
     /// </summary>
     public void ReleaseLock()
     {
-        ReleaseHandle();
+        // Contrary to intuition, this does not invalidate the handle (see comment on declaration)
+        Interop.libgpiod.gpiod_line_release(_handle);
     }
 
-    public override bool IsInvalid => handle == IntPtr.Zero || handle == LibgpiodV1.InvalidHandleValue;
+    public bool IsInvalid => _handle == IntPtr.Zero || _handle == Interop.libgpiod.InvalidHandleValue;
+
+    public void Dispose()
+    {
+        if (_handle != IntPtr.Zero)
+        {
+            Interop.libgpiod.gpiod_line_release(_handle);
+            _handle = IntPtr.Zero;
+        }
+    }
 }
